feat: compiles client daemon ring buffer, time for testing

+++ crates/nix-btm-common/src/daemon_side.rs
@@ -17,8 +17,8 @@ use crate::{
 
 /// Align `x` up to the next multiple of 2^`align_pow` bytes.
 #[inline]
-pub const fn align_up_pow2(num_bytes: u64, align_pow: u32) -> u64 {
-    let align = 1u64 << align_pow; // the actual alignment
+pub const fn align_up_pow2(num_bytes: u32, align_pow: u32) -> u32 {
+    let align = 1u32 << align_pow; // the actual alignment
     (num_bytes + (align - 1)/* round up */) & !(align - 1/* truncate down */)
 }
 
@@ -62,12 +62,12 @@ pub fn create_shmem_and_write_snapshot(
     // we need to calculate the size of the shit we're sending
     let state_blob = cbor::to_vec(&state_wire)?;
 
-    let header_len = size_of::<SnapshotHeader>() as u64;
-    let len_state_blob = state_blob.len() as u64;
+    let header_len = size_of::<SnapshotHeader>() as u32;
+    let len_state_blob = state_blob.len() as u32;
 
     // this is just in case the page size is < 4 required for the struct (how??)
     let total_len_struct_aligned =
-        align_up_pow2(header_len + len_state_blob, SC_HDR_ALIGN);
+        align_up_pow2(header_len + len_state_blob, SC_HDR_ALIGN) as u64;
     let total_len_snapshot = round_up_page(total_len_struct_aligned);
 
     let name = format!("nix-btm-snapshot-p{pid}");
@@ -84,7 +84,7 @@ pub fn create_shmem_and_write_snapshot(
     let mut mappedmem = unsafe { shmem.map(0x0)? };
     let buf = mappedmem.map();
 
-    let hdr = SnapshotHeader::new(len_state_blob, snap_seq_uid);
+    let hdr = SnapshotHeader::new(len_state_blob as u64, snap_seq_uid);
     let hdr_bytes = bytemuck::bytes_of(&hdr);
     buf[..hdr_bytes.len()].copy_from_slice(hdr_bytes);
     let start_state_blob = header_len as usize;
